requirements/run.txt
mmcv==2.1.0
mmdet
mmengine
ninja
pytorch3d
transformers
numba
tqdm
einops